pyproject.toml
[project]
name = "proyecto"
version = "0.1.0"
description = "Cargador de datos para imágenes térmicas y de profundidad"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "matplotlib>=3.7.0",
    "pillow>=11.2.1",
    "wandb>=0.19.11",
]

[project.scripts]
load-data = "src.main:main"
visualize-samples = "src.example:main"
train-model = "src.train:main"
visualize-results = "src.visualize:main"
